tg-etl/requirements.txt
PyYAML>=6.0
requests>=2.28.0
telethon>=1.28.0
configparser>=5.3.0
python-dotenv>=1.0.0
openai>=1.0.0 
pycountry